# requestly/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JobPilot API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/match/pending": {
      "get": {
        "tags": [
          "match"
        ],
        "summary": "Pending",
        "operationId": "pending_api_match_pending_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/match/selection/{job_id}": {
      "post": {
        "tags": [
          "match"
        ],
        "summary": "Record Selection",
        "description": "Resolves the FR-7.3 keyword interrupt. There is no bulk variant on purpose.",
        "operationId": "record_selection_api_match_selection__job_id__post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordSelection"
              }
            }
          }
        }
      }
    },
    "/api/match/writeMatch": {
      "post": {
        "tags": [
          "match"
        ],
        "summary": "Write Match",
        "operationId": "write_match_api_match_writeMatch_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchWrite"
              }
            }
          }
        }
      }
    },
    "/api/match/getMatch/{job_id}": {
      "get": {
        "tags": [
          "match"
        ],
        "summary": "Get Match",
        "operationId": "get_match_api_match_getMatch__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/resume/versions/{job_id}": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "List Versions",
        "operationId": "list_versions_api_resume_versions__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/resume/base": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Get Base Resume",
        "operationId": "get_base_resume_api_resume_base_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "resume"
        ],
        "summary": "Save Base Resume",
        "operationId": "save_base_resume_api_resume_base_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseResumeStore"
              }
            }
          }
        }
      }
    },
    "/api/resume/base/pdf": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Download Base Pdf",
        "description": "The stored .tex compiled, not re-rendered: the PDF has to be the document the\nuser is looking at, including anything the model rectified since the last render.",
        "operationId": "download_base_pdf_api_resume_base_pdf_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/resume/storeResume": {
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Store Resume",
        "operationId": "store_resume_api_resume_storeResume_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeStore"
              }
            }
          }
        }
      }
    },
    "/api/resume/getResume/{job_id}": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Get Resume",
        "operationId": "get_resume_api_resume_getResume__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/template/render/{template_id}": {
      "get": {
        "tags": [
          "template"
        ],
        "summary": "Render Template",
        "operationId": "render_template_api_template_render__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/template/download/{template_id}": {
      "get": {
        "tags": [
          "template"
        ],
        "summary": "Download Template",
        "description": "The same render as `/render`, handed over as the .tex itself. `/render` returns\nit inside JSON, where every newline is an escape the caller has to undo before\nthe source is usable.",
        "operationId": "download_template_api_template_download__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/template/preview/{template_id}": {
      "get": {
        "tags": [
          "template"
        ],
        "summary": "Template Preview",
        "operationId": "template_preview_api_template_preview__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/template": {
      "get": {
        "tags": [
          "template"
        ],
        "summary": "List Templates",
        "operationId": "list_templates_api_template_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/template/upload": {
      "post": {
        "tags": [
          "template"
        ],
        "summary": "Upload Template",
        "description": "The layout is read from the .tex itself, so the upload is the file and\nlittle else. A `%%! skills=grid location=city` line in the .tex overrides it.",
        "operationId": "upload_template_api_template_upload_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_template_api_template_upload_post"
              }
            }
          }
        }
      }
    },
    "/api/template/getTemplate/{template_id}": {
      "get": {
        "tags": [
          "template"
        ],
        "summary": "Get Template",
        "operationId": "get_template_api_template_getTemplate__template_id__get",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ]
      }
    },
    "/api/template/updateTemplate/{template_id}": {
      "patch": {
        "tags": [
          "template"
        ],
        "summary": "Update Template",
        "operationId": "update_template_api_template_updateTemplate__template_id__patch",
        "parameters": [
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PydanticObjectId"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateUpdate"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseResumeStore": {
        "properties": {
          "templateId": {
            "$ref": "#/components/schemas/PydanticObjectId"
          },
          "tex": {
            "type": "string",
            "minLength": 1,
            "title": "Tex"
          }
        },
        "type": "object",
        "required": [
          "templateId",
          "tex"
        ],
        "title": "BaseResumeStore",
        "description": "What the Resume screen submits once the user accepts the render."
      },
      "Body_upload_template_api_template_upload_post": {
        "properties": {
          "tex": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Tex",
            "description": ".tex source carrying {{TOKEN}} placeholders"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "preview": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview",
            "description": "optional PNG"
          }
        },
        "type": "object",
        "required": [
          "tex"
        ],
        "title": "Body_upload_template_api_template_upload_post"
      },
...
      "ContactStyle": {
        "type": "string",
        "enum": [
          "contactitem",
          "plain"
        ],
        "title": "ContactStyle"
      },
...
      "ExperienceStyle": {
        "type": "string",
        "enum": [
          "jobtitle",
          "timelinerole"
        ],
        "title": "ExperienceStyle"
      },
...
      "KeywordSelection": {
        "properties": {
          "selectedKeys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selectedkeys"
          },
          "skip": {
            "type": "boolean",
            "title": "Skip",
            "default": false
          }
        },
        "type": "object",
        "title": "KeywordSelection",
        "description": "The user's answer at the interrupt. An empty list is a valid answer."
      },
      "LineChange": {
        "properties": {
          "lineNo": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Lineno"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "previous": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous"
          }
        },
        "type": "object",
        "required": [
          "lineNo",
          "text"
        ],
        "title": "LineChange",
        "description": "One tailored line of the .tex, kept with what it replaced so the diff\nscreen and the source screen can never disagree."
      },
...
      "LocationStyle": {
        "type": "string",
        "enum": [
          "full",
          "city"
        ],
        "title": "LocationStyle"
      },
...
      "MatchWrite": {
        "properties": {
          "jobId": {
            "$ref": "#/components/schemas/PydanticObjectId"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "present": {
            "items": {
              "$ref": "#/components/schemas/PresentKeyword"
            },
            "type": "array",
            "title": "Present"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/MissingKeyword"
            },
            "type": "array",
            "title": "Missing"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskFlag"
            },
            "type": "array",
            "title": "Risks"
          },
          "modelName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelname"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "score"
        ],
        "title": "MatchWrite",
        "description": "What the matching agent posts. The JD vector goes to Atlas, never here."
      },
      "MissingKeyword": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "mentions": {
            "type": "integer",
            "title": "Mentions",
            "default": 1
          },
          "evidence": {
            "type": "string",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "evidence"
        ],
        "title": "MissingKeyword"
      },
      "PresentKeyword": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "chunkId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunkid"
          }
        },
        "type": "object",
        "required": [
          "label"
        ],
        "title": "PresentKeyword",
        "description": "Already in the resume. Read-only on the selection screen."
      },
...
      "PydanticObjectId": {
        "type": "string",
        "maxLength": 24,
        "minLength": 24,
        "pattern": "^[0-9a-f]{24}$",
        "example": "5eb7cf5a86d9755df3a6c593"
      },
...
      "ResumeStore": {
        "properties": {
          "jobId": {
            "$ref": "#/components/schemas/PydanticObjectId"
          },
          "filePath": {
            "type": "string",
            "title": "Filepath"
          },
          "templatePath": {
            "type": "string",
            "title": "Templatepath",
            "default": "templates/base_resume.tex"
          },
          "incorporated": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Incorporated"
          },
          "declined": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Declined"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/LineChange"
            },
            "type": "array",
            "title": "Changes"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "filePath"
        ],
        "title": "ResumeStore",
        "description": "What the tailoring agent posts once it has written the file."
      },
      "RiskFlag": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "key",
          "title",
          "detail"
        ],
        "title": "RiskFlag",
        "description": "FR-2.4 \u2014 surfaced, never selectable. Acting on one would be a misrepresentation."
      },
...
      "SkillsStyle": {
        "type": "string",
        "enum": [
          "grid",
          "pills"
        ],
        "title": "SkillsStyle"
      },
...
      "TemplateStyle": {
        "properties": {
          "contact": {
            "$ref": "#/components/schemas/ContactStyle",
            "default": "contactitem"
          },
          "skills": {
            "$ref": "#/components/schemas/SkillsStyle",
            "default": "pills"
          },
          "experience": {
            "$ref": "#/components/schemas/ExperienceStyle",
            "default": "jobtitle"
          },
          "location": {
            "$ref": "#/components/schemas/LocationStyle",
            "default": "full"
          },
          "columns": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Columns",
            "default": 4
          }
        },
        "type": "object",
        "title": "TemplateStyle",
        "description": "How this design lays out the blocks that differ between templates.\n\nAll of it except `location` is read off the uploaded .tex \u2014 see\n`service.infer_style`. A `%%!` directive line in the .tex overrides any of it."
      },
      "TemplateUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "style": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TemplateStyle"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "TemplateUpdate",
        "description": "Every field optional: a PATCH sends only what changed."
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "local \u2014 server/main.py"
    }
  ]
}